Finalize the move of existing modules to ix_apps package
