--- src/middlewared/middlewared/plugins/apps/schema.py
@@ -1,8 +1,0 @@
-from middlewared.service import Service
-
-
-class AppSchemaService(Service):
-
-    class Config:
-        namespace = 'app.schema'
-        private = True
